hide initiative type list when unique initiative type
