+++ decidim-initiatives/app/controllers/decidim/initiatives/admin/initiatives_controller.rb
@@ -8,6 +8,7 @@ module Decidim
       # Controller used to manage the initiatives
       class InitiativesController < Decidim::Initiatives::Admin::ApplicationController
         include Decidim::Initiatives::NeedsInitiative
+        include Decidim::Initiatives::SingleInitiativeType
         include Decidim::Initiatives::TypeSelectorOptions
         include Decidim::Initiatives::Admin::Filterable
