+++ decidim-initiatives/app/views/decidim/initiatives/admin/initiatives/_form.html.erb
@@ -22,6 +22,7 @@
     </div>
 
     <div class="row">
+      <% unless single_initiative_type? %>
       <div class="columns xlarge-6">
         <%= form.select :type_id,
                         initiative_type_options,
@@ -36,7 +37,7 @@
                           "data-signature-types-search-url": decidim_initiatives.initiative_type_signature_types_search_url
                         } %>
       </div>
-
+      <% end %>
       <div class="columns xlarge-6">
         <%= form.select :decidim_scope_id, [], {}, { disabled: !@form.signature_type_updatable? } %>
       </div>
